Move history/ddlmanager/timescale stuff around
